TerminalView: Move TextView logic that allows text selection into its own class, which allows scrolling selection.
